Refactor route to admin panels for manga CRUD operations, create manga controller and some views for this controller, create IMangaRepository interface for manga controller add repository pattern to manga controler with dependency injection
