+++ PudgeManga_Project/Models/Repositories/IAdminMangaRepository.cs
@@ -1,6 +1,6 @@
 ﻿namespace PudgeManga_Project.Models.Repositories
 {
-    public interface IRepository<T1, T2> where T1 : class
+    public interface IAdminMangaRepository<T1, T2> where T1 : class
     {
         Task<IEnumerable<T1>> GetAll();
         Task<T1> GetById(T2 id);
