+++ PudgeManga_Project/Views/AdminManga/Delete.cshtml
